refactor: extract shared HTTP transport into BaseResource
Committed-By-Agent: claude
